Design for Shopping Cart

+++ src/main/java/com/shoppingcart/model/ChargeRequest.java
@@ -10,7 +10,7 @@ public class ChargeRequest {
     }
 
     private String description;
-    private int amount; // cents
+    private double amount; // cents
     private Currency currency;
     private String stripeEmail;
     private String stripeToken;
